Reorganize yarn strategy module layout
