+++ src/Strategy/Yarn/V2/Resolvers.hs
@@ -9,7 +9,7 @@
 -- Resolvers
 --
 -- See also: default plugins, many of which contain resolvers https://github.com/yarnpkg/berry/blob/8afcaa2a954e196d6cd997f8ba506f776df83b1f/packages/yarnpkg-cli/package.json#L68-L82
-module Strategy.Yarn.Resolvers (
+module Strategy.Yarn.V2.Resolvers (
   Resolver (..),
   Package (..),
   resolveLocatorToPackage,
@@ -25,13 +25,16 @@ import Data.Text (Text)
 import Data.Text qualified as T
 import Data.Text.Extra (dropPrefix, showT)
 import Data.Void (Void)
-import Strategy.Yarn.LockfileV2
+import Strategy.Yarn.V2.Lockfile
 import Text.Megaparsec
 
 data Resolver = Resolver
-  { resolverName :: Text -- ^ Used for error messages
-  , resolverSupportsLocator :: Locator -> Bool -- ^ Does this resolver support the locator?
-  , resolverLocatorToPackage :: Locator -> Either Text Package -- ^ Convert this locator to a yarn package
+  { -- | Used for error messages
+    resolverName :: Text
+  , -- | Does this resolver support the locator?
+    resolverSupportsLocator :: Locator -> Bool
+  , -- | Convert this locator to a yarn package
+    resolverLocatorToPackage :: Locator -> Either Text Package
   }
 
 data Package
